Import of Michael Schlenker's enhancements to David's logger package.
Import of Michael Schlenker's bugfixes for David's logger
package.

Import Pat Thoyt's bugfix regarding setting of loglevels in
the dns module.

Import of Pat Thoyts bugfix regarding the installation of
the struct1 module.

Updated the version information in a number of packages to
sync with the numbers in the 1.6.1 release, or to pull ahead
and distinguish ourselves, where appropriate.


FossilOrigin-Name: 93369ff9aad41511fd4ac42542484cb36357ac41