Reimplemented cluster::sleep to avoid vwait. (Nested vwaits go out into space under the wrong conditions)
Added a new keyword "branchset" to the meta command. This allows you to input a dict, but tell
the system that all of the elements are leaf points. (The option keyword does that now, this is 
a more general purpose tool.)

FossilOrigin-Name: 3a80759025ba14ec76b2834f5a7aa0437775d010