Improvements to tool based on testing with IRM
All objects now get a "class" organ which points to their class

The "meta" method now defers to class if a value is not present in the
local "meta" dict

When an option of class "variable" is specified, tool now arranges to
create that variable on object startup

Bumped the version to reflect those improvements

The optimization that was only rebuilding ensemble methods for classes that
modified an ensemble wasn't working for complex class hierarchies. The new
rule is to rebuild ensembled methods for every new class defined or modified.
(Until something more efficient can be developed...)

FossilOrigin-Name: d04ac61c9bd1d919f5d7bff0cb429fb36d06a174