Bumped oo::meta to 0.5
Fixes two bugs. First: The dirty classes index was never being cleared, so 
the system never had a chance to work from cached data.

Second, the depth-first search used to track down ancestral data was getting
fooled by corner cases in the complex class hierarchies used in IRM. The system
has been re-jiggered so that an object only consults the metadata from the 
ancestors called out in the SUPERCLASS and MIXIN statements. (And the metadata
for those classes will be built as needed, and so on.)

FossilOrigin-Name: 49c8bc1070d309ed791d788a05671213822f8de2