Bumped oometa to version 0.7
Added a new hook to notify dialects that a class' metadata has changed

Bumped tool to version 0.6

Added a procedure to intercept calls from oometa's new "rebuild" command, and
signal that a class needs to rebuild it's ensemble methods.

Added a new family of procedures to be executed when a new tool object is instantiated.

On object startup, every object checks to see if they have a valid method ensemble.
If not, it triggers a rebuild. This replaces the prior scheme where ensembles were
built on response to tool::define

FossilOrigin-Name: 76c9fac30f619e3a3b2c46a112ed51cec24f023f