Fixed a bug in coroutine::util::gets. The prior blocking state was overwritten inside of the loop.
Added a new command to the coroutine::util namespace: gets_safety. It is intended for http servers which have to read in a line from a socket, but where attackers may attempt to pull off a stack overflow by sending a stream of data with no newline. 

Modified the http package to utilize the new gets_safety command.

FossilOrigin-Name: 1e347431489417e1fd260c1613762f01f4c66295