Redesigned the "not found" message inside of the httpd dispatcher to spawn off a reply object to handle the template processing for the error message.
Added handling for the special case of a browser communicating with localhost.
Modern browsers will not send a cookie for localhost, and in that case the
session id is marked as a constant string "localhost" to allow the rest
of the security system to operate properly.

FossilOrigin-Name: 1db0d3ebb869dafcf72c114027d98e641b4dccf86fd090658854c57701f15fd3