httpd module:
Replaced naked calls to [chan copy] with a new co-subroutine ChanCopy
which is part of the metaclass for the httpd module. Using ChanCopy means
the coroutine can remain in control for the entire process of processing
an http request. For large file transfers we were killing the coroutine
and waking the object back up with a fileevent.

Added a test for 404 Not Found errors

FossilOrigin-Name: bb27bb43b438da1648e6ac72b87c93774781af22d29b4e71e652355257cda3ea